# Use `check_logical()` rather than `vec_assert()` in `if_else()` and `vec_case_when()`