Closes https://github.com/tidyverse/dplyr/issues/6678

As mentioned in #6711, this also improves the performance of `case_when()` a little further:

``` r
library(dplyr)

# Size 1
x <- 1L

bench::mark(
  one = case_when(x == 1L ~ TRUE), 
  two = case_when(x == 1L ~ TRUE, x == 2L ~ NA),
  three = case_when(x == 1L ~ TRUE, x == 2L ~ NA, TRUE ~ FALSE),
  iterations = 50000,
  check = FALSE
)

# Main
#> # A tibble: 3 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 one           172µs    251µs     3851.   498.9KB     10.8
#> 2 two           212µs    268µs     3683.     1.3KB     13.6
#> 3 three         245µs    309µs     3214.     1.3KB     14.6

# This PR
#> # A tibble: 3 × 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 one           154µs    212µs     4614.   493.1KB     11.6
#> 2 two           180µs    229µs     4337.     1.3KB     13.4
#> 3 three         206µs    282µs     3277.     1.3KB     11.8
```